Include testing at compile time
