+++ dpc-attribution/pom.xml
@@ -26,7 +26,6 @@
             <groupId>gov.cms.dpc</groupId>
             <artifactId>dpc-testing</artifactId>
             <version>0.4.0-SNAPSHOT</version>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>io.dropwizard</groupId>
