+++ dpc-attribution/src/main/java/gov/cms/dpc/attribution/cli/SeedCommand.java
@@ -10,13 +10,10 @@ import gov.cms.dpc.attribution.dao.tables.records.OrganizationsRecord;
 import gov.cms.dpc.attribution.dao.tables.records.PatientsRecord;
 import gov.cms.dpc.attribution.dao.tables.records.ProvidersRecord;
 import gov.cms.dpc.attribution.jdbi.RosterUtils;
-import gov.cms.dpc.common.entities.AddressEntity;
-import gov.cms.dpc.common.entities.OrganizationEntity;
-import gov.cms.dpc.common.entities.PatientEntity;
-import gov.cms.dpc.common.entities.ProviderEntity;
-import gov.cms.dpc.common.utils.DBUtils;
+import gov.cms.dpc.common.entities.*;
 import gov.cms.dpc.common.utils.SeedProcessor;
 import gov.cms.dpc.fhir.converters.FHIREntityConverter;
+import gov.cms.dpc.testing.utils.DBUtils;
 import io.dropwizard.core.Application;
 import io.dropwizard.core.cli.EnvironmentCommand;
 import io.dropwizard.core.setup.Environment;
