+++ dpc-consent/pom.xml
@@ -90,7 +90,6 @@
             <groupId>gov.cms.dpc</groupId>
             <artifactId>dpc-testing</artifactId>
             <version>0.4.0-SNAPSHOT</version>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>com.h2database</groupId>
