+++ dpc-testing/src/main/java/gov/cms/dpc/testing/utils/DBUtils.java
@@ -1,6 +1,6 @@
-package gov.cms.dpc.common.utils;
+package gov.cms.dpc.testing.utils;
 
-import gov.cms.dpc.common.exceptions.NotATestEnvironmentException;
+import gov.cms.dpc.testing.exceptions.NotATestEnvironmentException;
 import org.jooq.DSLContext;
 import org.jooq.Schema;
 import org.slf4j.Logger;
@@ -35,7 +35,11 @@ public class DBUtils {
         }
 
         // Truncate all the tables (except for the liquibase metadata)
-        schemas.get(0).getTables().stream().filter(table -> !table.getName().startsWith("databasechangelog"))
+        schemas
+                .get(0)
+                .getTables()
+                .stream()
+                .filter(table -> !table.getName().startsWith("databasechangelog"))
                 .forEach(table -> {
                     logger.trace("Truncating table: {}", table.getName());
                     context.truncate(table).cascade().execute();
